requirements.txt
# Emotion Detection Dependencies
# Install with: pip install -r requirements.txt

deepface==0.0.89
opencv-python==4.9.0.80
numpy>=1.24.0,<2.0
tensorflow>=2.15.0
tf_keras>=2.15.0

# Optional: For serial communication with Arduino/hardware
pyserial==3.5
